refactor(protocol-designer): rename hook.ts to useKitchen.ts
rename hook.ts to useKitchen.ts

close AUTH-1985
