+++ protocol-designer/src/components/organisms/Ot2Modules/index.tsx
@@ -56,7 +56,7 @@ import { COMPATIBLE_LABWARE_ALLOWLIST_BY_MODULE_TYPE } from '../../../utils/labw
 import { FixedTrashText, MagnetModuleChangeContent } from '../../molecules'
 import { useBlockingHint } from '../BlockingHintModal'
 import { ConfirmDeleteEntityInUseModal } from '../ConfirmDeleteEntityInUseModal'
-import { useKitchen } from '../Kitchen/hooks'
+import { useKitchen } from '../Kitchen/useKitchen'
 import { ModuleEmptySelectorButtons } from '../ModuleEmptySelectorButtons'
 import { getNextAvailableModuleSlot, getSlotsWithCollisions } from '../utils'
 import { getModuleOnSlot } from './util'
